Split mercator parts out of terrarium for re-use.
